Separate contact_models and contact_policies.
