+++ src/sid/rapid_tests.py
@@ -6,7 +6,6 @@ from typing import Optional
 
 import numpy as np
 import pandas as pd
-from sid.config import DTYPE_N_CONTACTS
 from sid.shared import boolean_choices
 from sid.shared import separate_contact_model_names
 from sid.validation import validate_return_is_series_or_ndarray
@@ -35,12 +34,12 @@ def perform_rapid_tests(
         contacts = pd.concat([recurrent_contacts, random_contacts], axis=1)
 
         receives_rapid_test = _compute_who_receives_rapid_tests(
-            date,
-            states,
-            params,
-            rapid_test_models,
-            contacts,
-            seed,
+            date=date,
+            states=states,
+            params=params,
+            rapid_test_models=rapid_test_models,
+            contacts=contacts,
+            seed=seed,
         )
 
         is_tested_positive = _sample_test_outcome(
@@ -68,12 +67,6 @@ def apply_reactions_to_rapid_tests(
     if rapid_test_reaction_models:
         recurrent_names, random_names = separate_contact_model_names(contact_models)
 
-        recurrent_contacts = pd.DataFrame(
-            recurrent_contacts, index=states.index, columns=recurrent_names
-        )
-        random_contacts = pd.DataFrame(
-            random_contacts, index=states.index, columns=random_names
-        )
         contacts = pd.concat([recurrent_contacts, random_contacts], axis=1)
 
         for model in rapid_test_reaction_models.values():
@@ -88,8 +81,8 @@ def apply_reactions_to_rapid_tests(
                     seed=next(seed),
                 )
 
-        recurrent_contacts = contacts[recurrent_names].to_numpy(dtype=bool)
-        random_contacts = contacts[random_names].to_numpy(dtype=DTYPE_N_CONTACTS)
+        recurrent_contacts = contacts[recurrent_names]
+        random_contacts = contacts[random_names]
 
     return recurrent_contacts, random_contacts
